requirements.txt
Django==1.8.9
django-bootstrap-pagination==1.6.1
django-bootstrap3==6.2.2
django-embed-video==1.1.0
django-flat-theme==1.1.3
django-registration-redux==1.3
django-thumbs==0.4
flup==1.0.2
mysql-python==1.2.5
Pillow==3.1.0
requests==2.9.1
slugify==0.0.1
wheel==0.26.0




